add CLI postMessage for token
