+++ packages/app/src/app/pages/CLI/elements.ts
@@ -11,3 +11,18 @@ export const Container = styled.div`
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
     Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue';
 `;
+
+export const ContentContainer = styled.div`
+  height: 100%;
+  width: 100%;
+  text-align: left;
+`;
+
+export const Buttons = styled.div`
+  display: flex;
+  justify-content: center;
+
+  > button {
+    margin: 1rem;
+  }
+`;
